feat: refactor initramfs and initcall handling for improved task management
